+++ tx-backend/src/main/java/org/eclipse/tractusx/traceability/infrastructure/edc/blackbox/asset/Asset.java
@@ -27,6 +27,8 @@ import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import com.fasterxml.jackson.databind.annotation.JsonPOJOBuilder;
 import lombok.ToString;
 import lombok.extern.slf4j.Slf4j;
+import org.eclipse.tractusx.traceability.qualitynotification.domain.model.QualityNotificationMessage;
+import org.eclipse.tractusx.traceability.qualitynotification.domain.model.QualityNotificationType;
 
 import java.util.HashMap;
 import java.util.Map;
@@ -46,6 +48,7 @@ public class Asset {
     private static final String ASSET_VALUE_NOTIFICATION_METHOD_RECEIVE = "receive";
     public static final String ASSET_KEY_NOTIFICATION_TYPE = "asset:prop:notificationtype";
     public static final String ASSET_VALUE_QUALITY_INVESTIGATION = "qualityinvestigation";
+    public static final String ASSET_VALUE_QUALITY_ALERT = "qualityalert";
     public static final String PROPERTY_ID = "asset:prop:id";
     public static final String PROPERTY_NAME = "asset:prop:name";
     public static final String PROPERTY_DESCRIPTION = "asset:prop:description";
@@ -96,13 +99,10 @@ public class Asset {
     }
 
     @JsonIgnore
-    public boolean isQualityInvestigationReceive() {
-        return ASSET_VALUE_QUALITY_INVESTIGATION.equals(this.getPropertyNotificationType()) && ASSET_VALUE_NOTIFICATION_METHOD_RECEIVE.equals(this.getPropertyNotificationMethod());
-    }
-
-    @JsonIgnore
-    public boolean isQualityInvestigationUpdate() {
-        return ASSET_VALUE_QUALITY_INVESTIGATION.equals(this.getPropertyNotificationType()) && ASSET_VALUE_NOTIFICATION_METHOD_UPDATE.equals(this.getPropertyNotificationMethod());
+    public boolean isQualityNotificationOffer(QualityNotificationMessage qualityNotificationMessage) {
+        final String propertyNotificationTypeValue = QualityNotificationType.ALERT.equals(qualityNotificationMessage.getType()) ? ASSET_VALUE_QUALITY_ALERT : ASSET_VALUE_QUALITY_INVESTIGATION;
+        final String propertyMethodValue = qualityNotificationMessage.getIsInitial() ? ASSET_VALUE_NOTIFICATION_METHOD_RECEIVE : ASSET_VALUE_NOTIFICATION_METHOD_UPDATE;
+        return propertyNotificationTypeValue.equals(this.getPropertyNotificationType()) && propertyMethodValue.equals(this.getPropertyNotificationMethod());
     }
 
     public Map<String, Object> getProperties() {