+++ tx-backend/src/main/java/org/eclipse/tractusx/traceability/qualitynotification/application/investigation/service/InvestigationService.java
@@ -19,7 +19,13 @@
 package org.eclipse.tractusx.traceability.qualitynotification.application.investigation.service;
 
 import org.eclipse.tractusx.traceability.qualitynotification.application.service.QualityNotificationService;
+import org.eclipse.tractusx.traceability.qualitynotification.domain.model.QualityNotificationId;
+import org.eclipse.tractusx.traceability.qualitynotification.domain.model.QualityNotificationSeverity;
+
+import java.time.Instant;
+import java.util.List;
 
 public interface InvestigationService extends QualityNotificationService {
+    QualityNotificationId start(List<String> partIds, String description, Instant targetDate, QualityNotificationSeverity severity);
 
 }