+++ tx-backend/src/main/java/org/eclipse/tractusx/traceability/qualitynotification/application/service/QualityNotificationService.java
@@ -21,15 +21,10 @@ package org.eclipse.tractusx.traceability.qualitynotification.application.servic
 import org.eclipse.tractusx.traceability.common.model.PageResult;
 import org.eclipse.tractusx.traceability.qualitynotification.domain.model.QualityNotification;
 import org.eclipse.tractusx.traceability.qualitynotification.domain.model.QualityNotificationId;
-import org.eclipse.tractusx.traceability.qualitynotification.domain.model.QualityNotificationSeverity;
 import org.eclipse.tractusx.traceability.qualitynotification.domain.model.QualityNotificationStatus;
 import org.springframework.data.domain.Pageable;
 
-import java.time.Instant;
-import java.util.List;
-
 public interface QualityNotificationService {
-    QualityNotificationId start(List<String> partIds, String description, Instant targetDate, QualityNotificationSeverity severity);
 
     PageResult<QualityNotification> getCreated(Pageable pageable);