+++ tx-backend/src/main/java/org/eclipse/tractusx/traceability/qualitynotification/domain/model/QualityNotificationType.java
@@ -1,7 +1,5 @@
 /********************************************************************************
- * Copyright (c) 2022, 2023 Bayerische Motoren Werke Aktiengesellschaft (BMW AG)
- * Copyright (c) 2022, 2023 ZF Friedrichshafen AG
- * Copyright (c) 2022, 2023 Contributors to the Eclipse Foundation
+ * Copyright (c) 2023 Contributors to the Eclipse Foundation
  *
  * See the NOTICE file(s) distributed with this work for additional
  * information regarding copyright ownership.
@@ -19,13 +17,9 @@
  * SPDX-License-Identifier: Apache-2.0
  ********************************************************************************/
 
-package org.eclipse.tractusx.traceability.qualitynotification.infrastructure.investigation.repository;
+package org.eclipse.tractusx.traceability.qualitynotification.domain.model;
 
-import org.eclipse.tractusx.traceability.qualitynotification.infrastructure.investigation.model.InvestigationNotificationEntity;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface JpaNotificationRepository extends JpaRepository<InvestigationNotificationEntity, String> {
-
+public enum QualityNotificationType {
+    INVESTIGATION,
+    ALERT
 }