+++ tx-cucumber-tests/src/test/java/org/eclipse/tractusx/traceability/test/tooling/rest/Authentication.java
@@ -24,7 +24,6 @@ import io.restassured.http.Header;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
-import lombok.extern.slf4j.Slf4j;
 import org.eclipse.tractusx.traceability.test.tooling.EnvVariablesResolver;
 
 import java.util.HashMap;